+++ management/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.21.1 (Unreleased)
+
+* **Enhancement** Added `SloWindow` optional attribute to the Application SAML objects.
+* **Enhancement** Added SLO optional attributes to the SAML External Identity Provider object.
+
 # v0.21.0 (2023-05-22)
 
 * **Note** Deprecated `bundleId` and `packageName` at the root level of the `ApplicationOIDC` model. Customers should use `mobile.bundleId` and `mobile.packageName` going forward. [#172](https://github.com/patrickcping/pingone-go-sdk-v2/pull/172)